requirements.txt
gunicorn==20.1.0
flask==2.3.2
supabase==2.3.0
python-dotenv==1.0.0
qrcode[pillow]==7.4.2
validators==0.20.0